requirements.txt
django
djangorestframework
requests
beautifulsoup4
selenium
python-dotenv
webdriver-manager
gunicorn
